Move transaction related components into their own subdirectory
